+++ streamlit/component-template/template-vite/vite_vanilla_component/frontend/main.js
@@ -1,7 +1,8 @@
 import { Streamlit } from "streamlit-component-lib"
 import G6 from '@antv/g6'
 
-let selectedNode = null
+let selectedNodes = []
+let selectedEdges = []
 let graph = null
 
 const span = document.body.appendChild(document.createElement("span"))
@@ -15,11 +16,11 @@ button.onfocus = function() { isFocused = true }
 button.onblur = function() { isFocused = false }
 button.onclick = function() {
   numClicks += 1
-  Streamlit.setComponentValue({ numClicks, selectedNode })
+  Streamlit.setComponentValue({ numClicks, selectedNodes, selectedEdges })
 }
 
 function onRender(event) {
-  console.log('RENDER EVENT', Date.now()) //  current status... of numClicks and selectedNode...
+  console.log('RENDER EVENT', Date.now())
   const data = event.detail
   if (data.theme) {
     const borderStyling = `1px solid var(${
@@ -32,6 +33,11 @@ function onRender(event) {
   let name = data.args["name"]
   textNode.textContent = `Hello, ${name}! ` + String.fromCharCode(160)
 
+  const options = data.args["options"] || {
+    selectNodes: 3, // 0, 1, n
+    selectEdges: 3, //0, 1, n
+    clickClearAll: false, // clear all nodes if clicking on canvas
+  }
   let nodes = data.args["nodes"]
   let edges = data.args["edges"]
   const gdata = {
@@ -47,34 +53,79 @@ function onRender(event) {
   container.setAttribute('id', 'container');
 
   if (!graph) {
-    config.width = document.getElementById('container').scrollWidth;
-    config.height = document.getElementById('container').scrollHeight || 500;
+    config.width = config.width || document.getElementById('container').scrollWidth;
+    config.height = config.height || document.getElementById('container').scrollHeight || 500;
   
     graph = new G6.Graph(config)
 
     // graph events handling
     graph.on('edge:mouseenter', ({item}) => graph.setItemState(item, 'active', true))
     graph.on('edge:mouseleave', ({item}) => graph.setItemState(item, 'active', false))
-    graph.on('edge:click', ({item}) => graph.setItemState(item, 'selected', true))
-    graph.on('node:click', (evt) => {
-      const { item } = evt
-      graph.getNodes().forEach((node) => graph.clearItemStates(node))
+
+    graph.on('edge:click', ({item}) => {
       graph.setItemState(item, 'selected', true)
-      const { id, label } =  item._cfg.model
-      selectedNode = { id, label } // ${id} ${label}`
+      if (options.selectEdges) {
+        const selected = item._cfg.states.includes('selected')
+        if (options.selectEdges > 1) { // multiple
+          graph.setItemState(item, 'selected', selected ? false : true)
+          if (selected) { // remove
+            selectedEdges = selectedEdges.filter((node) => node.id !== item._cfg.model.id)
+          } else { // add
+            selectedEdges.push(item._cfg.model)
+          }
+        } else { // single
+          graph.getNodes().forEach((node) => graph.clearItemStates(node)) // unselect the rest
+          selectedEdges = []
+          if (!selected) {
+            graph.setItemState(item, 'selected', true) // select this node
+            selectedEdges = [item._cfg.model] 
+          }
+        }
+      }
     })
+
+    graph.on('node:click', ({item}) => {
+      if (options.selectNodes) {
+        const selected = item._cfg.states.includes('selected')
+        if (options.selectNodes > 1) { // multiple
+          graph.setItemState(item, 'selected', selected ? false : true)
+          if (selected) { // remove
+            selectedNodes = selectedNodes.filter((node) => node.id !== item._cfg.model.id)
+          } else { // add
+            selectedNodes.push(item._cfg.model)
+          }
+        } else { // single
+          graph.getNodes().forEach((node) => graph.clearItemStates(node)) // unselect the rest
+          selectedNodes = []
+          if (!selected) {
+            graph.setItemState(item, 'selected', true) // select this node
+            selectedNodes = [item._cfg.model] 
+          }
+        }
+      }
+    })
+
     graph.on('canvas:click', (evt) => {
-      selectedNode = null
-      graph.getEdges().forEach((edge) => graph.clearItemStates(edge))
-      graph.getNodes().forEach((node) => graph.clearItemStates(node))
+      if (options.clickClearAll) {
+        selectedNodes = []
+        selectedEdges = []
+        graph.getEdges().forEach((edge) => graph.clearItemStates(edge))
+        graph.getNodes().forEach((node) => graph.clearItemStates(node))  
+      }
     })
   }
   graph.data(gdata);
   graph.render();
-  if (selectedNode) { // highlight node based on label
-    const foundNode = graph.getNodes().find((node) => node._cfg.model.label === selectedNode.label)
+
+  selectedNodes.forEach((selectedNode) => { // highlight nodes
+    const foundNode = graph.getNodes().find((_node) => _node._cfg.model.label === selectedNode.label)
     if (foundNode) graph.setItemState(foundNode, 'selected', true)
-  }
+  })
+
+  selectedEdges.forEach((selectedEdge) => { // highlight edges
+    const foundNode = graph.getNodes().find((_node) => _node._cfg.model.label === selectedEdge.label)
+    if (foundNode) graph.setItemState(foundNode, 'selected', true)
+  })
 
   Streamlit.setFrameHeight()
 }
@@ -82,3 +133,8 @@ function onRender(event) {
 Streamlit.events.addEventListener(Streamlit.RENDER_EVENT, onRender)
 Streamlit.setComponentReady()
 Streamlit.setFrameHeight()
+
+// IN PROGRESS keep state of expanders using session state,  keep the result of the calculate after each submission
+// IN PROGRESS app session metadata error on first load... to check
+// TBD Server side caching? limited memory - cache data at server... (check on memory usage and limitations)
+// TBD any multiple tabs available on streamlit? able to have expander only show 1 at a time?